Stop referencing unused properties
A few fields for dmaap object were still being set assuming that values
came from properties file.  Furthermore, a private variable in the
DmaapService class was not set to  new value from request.
Result was that lack of property caused topicNsRoot field to be null,
and then the null was used to compose AAF action/instance values incorrectly.

Change-Id: Iffabf5728fdfb2f9496c7eb248cb1ed9e1cc1536
Signed-off-by: dglFromAtt <dgl@research.att.com>
Issue-ID: DMAAP-933