Update patch set 1
Patch Set 1: Code-Review+2

The enablement of authentication is a bit disjoint right now.  My direction to Olaf was to have the code default to the way we desire it to be (enabled).  This is easily controlled through dmaap-bc properties, which may be disabled for deployment until all the components are working.  Also, we are investigating a more global (cross component) method for enabling/disabling this. More to come on this subject.  Meanwhile I will merge this.

Patch-set: 1
Label: Code-Review=+2