Update patch set 1
Patch Set 1:

The code on some of these projects is a mess.
Problem here is most of the code uses the BaseLoggingClass to log to EELF, but some stuff still uses log4j.

To determine which ones, I removed the log4j 1 jar from the project then tried to build it in the IDE.
Most of the classes are in the AAF module.

Also, the following version need to be updated.

jetty version - 9.4.36.v20210114

Patch-set: 1
Reviewer: Gerrit User 4405 <4405@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>